+++ app/src/main/java/com/test/weather/feature/forecast/response/Day.kt
@@ -9,21 +9,21 @@ data class Day(@SerializedName("avgvis_km")
                @SerializedName("avgtemp_f")
                val avgtempF: Double = 0.0,
                @SerializedName("avgtemp_c")
-               val avgtempC: Int = 0,
+               val avgtempC: Double = 0.0,
                @SerializedName("maxtemp_c")
                val maxtempC: Double = 0.0,
                @SerializedName("maxtemp_f")
-               val maxtempF: Int = 0,
+               val maxtempF: Double = 0.0,
                @SerializedName("mintemp_c")
                val mintempC: Double = 0.0,
                @SerializedName("avgvis_miles")
-               val avgvisMiles: Int = 0,
+               val avgvisMiles: Double = 0.0,
                @SerializedName("mintemp_f")
                val mintempF: Double = 0.0,
                @SerializedName("totalprecip_in")
                val totalprecipIn: Double = 0.0,
                @SerializedName("avghumidity")
-               val avghumidity: Int = 0,
+               val avghumidity: Double = 0.0,
                @SerializedName("condition")
                val condition: Condition,
                @SerializedName("maxwind_kph")
